Move compat shim in Standard Provider to `version_compat.py`
Moves the conditional imports to `version_compat.py`
